Revert "feat(hslink-ob): add hslink ob project"
This reverts commit 672f85412119da4e9380f26c3874d7b539bb1623.
