--- projects/HSLink-OB/.version
@@ -1,1 +1,0 @@
-2.4.2
