--- projects/HSLink-OB/dp_common.h
@@ -1,13 +1,0 @@
-//
-// Created by HalfSweet on 25-3-23.
-//
-
-#ifndef HSLINK_PRO_DP_COMMON_H
-#define HSLINK_PRO_DP_COMMON_H
-
-#include <stdint.h>
-
-uint8_t software_reset(void);
-void por_reset(void);
-
-#endif //HSLINK_PRO_DP_COMMON_H
