+++ projects/hpm5301evklite/.gitignore
@@ -1,0 +1,28 @@
+.DS_Store
+.swp
+.bak
+cscope.*
+*.orig
+*.patch
+
+build
+*_build
+cmake-build*
+.idea
+.vscode
+.cache
+
+*.log
+*.jou
+.settings
+.cproject
+.project
+
+doctrees
+doc/output/*
+
+segger_embedded_studio
+*.pyc
+sec_core_img.c
+*.o
+
