+++ projects/hpm5301evklite/app.yaml
@@ -1,2 +1,2 @@
 minimum_sdk_version:
-  - 1.8.0
+  - 1.6.0
